requirements.txt
ultralytics>=8.0.0
opencv-python>=4.5.0
numpy>=1.20.0
pandas>=1.3.0
matplotlib>=3.4.0
scikit-learn>=0.24.0
python-dotenv>=0.19.0
simple-salesforce>=1.10.0
slack-bolt>=1.9.0
web3>=5.20.0
joblib>=1.0.0
requests>=2.27.0
pillow>=9.0.0
geopy>=2.2.0
googlemaps>=4.6.0
python-dateutil>=2.8.2
